use tsup to emit ts lib
