+++ peer/index.html
@@ -28,7 +28,7 @@
     </form>
   </nav>
 
-  <script type="module" src="/src/main.ts"></script>
+  <script type="module" src="/main.ts"></script>
 </body>
 
 </html>
