+++ peer/main.ts
@@ -1,4 +1,4 @@
-import { createPeer, PeerStore } from "./lib.ts";
+import { createPeer, PeerStore } from "./src/index.ts";
 
 (async () => {
   const urlParams = new URLSearchParams(window.location.search);
@@ -17,7 +17,7 @@ import { createPeer, PeerStore } from "./lib.ts";
     baseUrl: baseUrl || undefined,
     token,
   });
-  const portal = new PeerStore(peer);
+  const store = new PeerStore(peer);
 
   const textDOM = document.getElementById("text")! as HTMLInputElement;
   const formDOM = document.getElementById("form")! as HTMLFormElement;
@@ -27,13 +27,13 @@ import { createPeer, PeerStore } from "./lib.ts";
     e.preventDefault();
 
     const key = `${new Date().toISOString()} (${peerId})`;
-    portal.$store.setKey(key, textDOM.value);
+    store.$kv.setKey(key, textDOM.value);
     textDOM.value = "";
   };
 
-  portal.$store.listen(() => {
+  store.$kv.listen(() => {
     // TODO: this is naive, this can be done more efficiently with a binary heap or bst.
-    const snapshot = portal.$store.get();
+    const snapshot = store.$kv.get();
     const messages = Object.entries(snapshot);
 
     messages.sort((a, b) => a[0].localeCompare(b[0]));
@@ -44,5 +44,5 @@ import { createPeer, PeerStore } from "./lib.ts";
       containerDOM.appendChild(msgDOM);
     }
   });
-  portal.start();
+  store.start();
 })();
