+++ peer/package.json
@@ -5,18 +5,23 @@
   "files": [
     "dist"
   ],
-  "main": "./dist/peer.umd.cjs",
-  "module": "./dist/peer.js",
+  "main": "./dist/index.cjs",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.mts",
   "exports": {
-    ".": {
-      "import": "./dist/peer.js",
-      "require": "./dist/peer.umd.cjs"
+    "import": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "require": {
+      "types": "./dist/index.d.cts",
+      "require": "./dist/index.cjs"
     }
   },
   "scripts": {
     "gen": "protoc --ts_out=. --experimental_allow_proto3_optional --ts_opt=client_generic --proto_path proto/v1 ./proto/v1/signaling.proto && node ./scripts/patch.cjs",
     "dev": "vite",
-    "build": "vite build",
+    "build": "tsup src/index.ts --format esm,cjs --dts --minify",
     "clean": "rm -rf dist",
     "test": "vitest",
     "preview": "vite preview",
@@ -46,10 +51,9 @@
     "@protobuf-ts/plugin": "^2.9.4",
     "jsr": "^0.13.2",
     "protoc-gen-ts": "^0.8.7",
-    "tsup": "^8.3.6",
+    "tsup": "^8.4.0",
     "typescript": "^5.7.2",
     "vite": "^6.2.0",
-    "vite-plugin-dts": "^4.5.3",
     "vitest": "^3.0.5",
     "webrtc-adapter": "^9.0.1"
   }
