+++ peer/vite.config.ts
@@ -1,27 +1,8 @@
-import { dirname, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
 import { defineConfig } from "vite";
 import { auth } from "@pulsebeam/demo-server";
-import dts from "vite-plugin-dts";
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export default defineConfig({
   base: "./",
-  plugins: [dts()],
-  build: {
-    lib: {
-      entry: resolve(__dirname, "src/lib.ts"),
-      name: "peer",
-      // the proper extensions will be added
-      fileName: "peer",
-    },
-    // rollupOptions: {
-    //   input: {
-    //     main: resolve(__dirname, "index.html"),
-    //   },
-    // },
-  },
   server: {
     proxy: {
       "/auth": {
